chore: remove unnecessary module(preview)
